pyproject.toml
[project]
name = "langchain_rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dashscope>=1.23.9",
    "faiss-cpu>=1.11.0.post1",
    "jupyter>=1.1.1",
    "pypdf2>=3.0.1",
    "streamlit>=1.47.0",
    "langchain>=0.3.0",
    "langchain-community>=0.3.0",
    "langchain-openai>=0.3.0",
    "python-dotenv>=1.0.0",
    "langchain-core>=0.3.0",
]
